pyproject.toml
[tool.poetry]
name = "telegram-gpt-bot"
version = "0.1.0"
description = "Telegram bot using ChatGPT"
authors = ["Canarypwn"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.6"
openai = "^1.50.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
